+++ apps/server/functions/package.json
@@ -1,6 +1,6 @@
 {
   "name": "server-functions",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "type": "module",
   "scripts": {
     "build": "sh ./.scripts/build.sh",
